handle_done_commits_*(): commits_before_their_time : move elements directly w/o temporary vector

+++ src/xwayland_ctx.hpp
@@ -40,7 +40,7 @@ struct focus_t
 struct CommitDoneEntry_t
 {
 	uint32_t winSeq;
-	uint32_t commitID;
+	uint64_t commitID;
 	uint64_t desiredPresentTime;
 	bool fifo;
 	uint16_t presentMargin;
